+++ src/utils.ts
@@ -3,21 +3,23 @@ import { create as createMultihash } from 'multiformats/hashes/digest'
 
 const KECCAK_256_CODE = 0x1b
 const ETH_TX_CODE = 0x93
+
 export class Utils {
   /**
-   * Flatten array of arrays
-   * @param arr - Array of arrays
-   */
-  static flattenArray(arr: any[]): any[] {
-    return arr.reduce((accumulator, value) => accumulator.concat(value), [])
-  }
-
-  /**
    * "sleeps" for the given number of milliseconds
-   * @param mills
    */
-  static async delay(mills: number): Promise<void> {
-    await new Promise<void>((resolve) => setTimeout(() => resolve(), mills))
+  static delay(mills: number, abortSignal?: AbortSignal): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+      const timeout = setTimeout(resolve, mills)
+      if (abortSignal) {
+        const done = () => {
+          clearTimeout(timeout)
+          reject(new Error(`Delay aborted`))
+        }
+        if (abortSignal.aborted) done()
+        abortSignal.addEventListener('abort', done)
+      }
+    })
   }
 
   /**
@@ -36,10 +38,9 @@ export class Utils {
    * @param arr - Array of number
    */
   static averageArray(arr: number[]): number {
-
     if (arr.length == 0) {
-        return 0
-    }    
+      return 0
+    }
 
     return arr.reduce((total, aNumber) => total + aNumber, 0) / arr.length
   }
